Add files via upload
The purpose of this analysis is to completely translate the message from a multi-lingual text to a word or phrase translation and subsequently, they examine the level of toxicity of the message or statement. The test was performed after the message was made available in multiple languages or it was determined that the message was invalid or uninteresting. Talking on social networks is often abusive and abusive. Team members often want to make sure that the discussion will not be too serious by using this service.